Moved key_split() to utils.py
